Move mirror discovery out of main.lua
autoDiscoverAndSetBaseUrl was 346 lines -- a fifth of main.lua, and its largest
function by half again. It touched one field of the plugin instance,
self.discover_channel, and called no method on it but itself: a feature that
happened to live on the plugin object rather than part of one.

It is now zlibrary/discovery.lua. main.lua keeps a five-line method delegating
to it, because the body recurses through that method and Ui calls it on the
plugin instance when offering auto-discovery after a failure.

The move is verifiable rather than merely careful. Discovery.run takes a
parameter literally named `self`, so the body is byte-for-byte what it was --
343 lines, checked against git rather than eyeballed in a diff. A move that
changes no line cannot change behaviour, which was worth more here than a
tidier signature.

main.lua: 1794 -> 1455 lines. No translatable string moved, gained or lost.

The harness checks what a wrapper gets wrong -- a dropped return, a lost
argument, a missing require -- and finds the names the module uses rather than
listing them, since the failure worth guarding against is the dependency nobody
remembered was there. It caught one during writing: `local a, b` with no
assignment is a binding, and the first version reported connection_menu as
unbound.

What is left in main.lua is coupled and stays: showMyBooksDialog calls eight
sibling methods, displaySearchResults owns six fields of search state. Those
are orchestration. downloadBook at 244 lines is the next candidate, and a
larger one than this.
